Move logic for corrected_info to a separate class
